fix: bypass_llm for Phase A + fp32 guards for linalg ops
- HierarchicalMultiHopModel.bypass_llm skips TopoBridge entirely
- Phase A sets bypass_llm=True, moves TopoBridge to CPU (5.6 vs 15.3 GB VRAM)
- Phase B/C re-enables TopoBridge on GPU
- All torch.linalg.eigh/svd/lstsq calls ensure fp32 input (AMP compat)
- AMP disabled: spectral-dominated model doesn't benefit from fp16
- Eliminates NaN grads from gate-spike TopoBridge noise in Phase A

+++ src/benchmarks/spectral_tasks.py
@@ -217,7 +217,7 @@ def generate_hodge_class_task(
         L1 = B1.T @ B1
         if B2 is not None:
             L1 = L1 + B2 @ B2.T
-        eigenvalues, eigenvectors = torch.linalg.eigh(L1)
+        eigenvalues, eigenvectors = torch.linalg.eigh(L1.float())
         harmonic_mask = eigenvalues.abs() < 1e-5
         if harmonic_mask.sum() > 0:
             harmonic_basis = eigenvectors[:, harmonic_mask]
